chore: refactor dir structure
