+++ lib/macos_ui.dart
@@ -37,8 +37,9 @@ export 'src/layout/content_area.dart';
 export 'src/layout/macos_list_tile.dart';
 export 'src/layout/resizable_pane.dart';
 export 'src/layout/scaffold.dart';
-export 'src/layout/sidebar.dart';
-export 'src/layout/sidebar_item.dart';
+export 'src/layout/sidebar/sidebar.dart';
+export 'src/layout/sidebar/sidebar_item.dart';
+export 'src/layout/sidebar/sidebar_items.dart';
 export 'src/layout/title_bar.dart';
 export 'src/layout/toolbar/custom_toolbar_item.dart';
 export 'src/layout/toolbar/toolbar.dart';
